+++ packages/react-router/lib/dom/ssr/components.tsx
@@ -227,6 +227,12 @@ export interface LinksProps {
    * element
    */
   nonce?: string | undefined;
+  /**
+   * A [`crossOrigin`](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/crossorigin)
+   * attribute to render on the [`<link>`](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/link)
+   * element
+   */
+  crossOrigin?: "anonymous" | "use-credentials";
 }
 
 /**
@@ -254,10 +260,11 @@ export interface LinksProps {
  * @mode framework
  * @param props Props
  * @param {LinksProps.nonce} props.nonce n/a
+ * @param {LinksProps.crossOrigin} props.crossOrigin n/a
  * @returns A collection of React elements for [`<link>`](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/link)
  * tags
  */
-export function Links({ nonce }: LinksProps): React.JSX.Element {
+export function Links({ nonce, crossOrigin }: LinksProps): React.JSX.Element {
   let { isSpaMode, manifest, routeModules, criticalCss } =
     useFrameworkContext();
   let { errors, matches: routerMatches } = useDataRouterStateContext();
@@ -274,6 +281,7 @@ export function Links({ nonce }: LinksProps): React.JSX.Element {
       {typeof criticalCss === "string" ? (
         <style
           {...{ [CRITICAL_CSS_DATA_ATTRIBUTE]: "" }}
+          nonce={nonce}
           dangerouslySetInnerHTML={{ __html: criticalCss }}
         />
       ) : null}
@@ -283,13 +291,24 @@ export function Links({ nonce }: LinksProps): React.JSX.Element {
           rel="stylesheet"
           href={criticalCss.href}
           nonce={nonce}
+          crossOrigin={crossOrigin}
         />
       ) : null}
       {keyedLinks.map(({ key, link }) =>
         isPageLinkDescriptor(link) ? (
-          <PrefetchPageLinks key={key} nonce={nonce} {...link} />
+          <PrefetchPageLinks
+            key={key}
+            nonce={nonce}
+            {...link}
+            crossOrigin={link.crossOrigin ?? crossOrigin}
+          />
         ) : (
-          <link key={key} nonce={nonce} {...link} />
+          <link
+            key={key}
+            nonce={nonce}
+            {...link}
+            crossOrigin={link.crossOrigin ?? crossOrigin}
+          />
         ),
       )}
     </>
@@ -487,7 +506,12 @@ function PrefetchPageLinksImpl({
       {keyedPrefetchLinks.map(({ key, link }) => (
         // these don't spread `linkProps` because they are full link descriptors
         // already with their own props
-        <link key={key} nonce={linkProps.nonce} {...link} />
+        <link
+          key={key}
+          nonce={linkProps.nonce}
+          {...link}
+          crossOrigin={link.crossOrigin ?? linkProps.crossOrigin}
+        />
       ))}
     </>
   );